lib/lambda/cnvFile/test_env/pyproject.toml
[tool.poetry]
name = "cnvfile"
version = "0.1.0"
description = ""
authors = ["admiswalker <admiswalker@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.4.0"
numpy = "^1.21.1"
urllib3 = "^1.26.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
